Split in POJO and use Argument
