+++ configurations/liquibase/src/main/java/io/micronaut/dbmigration/liquibase/management/endpoint/LiquibaseEndpoint.java
@@ -1,12 +1,25 @@
+/*
+ * Copyright 2017-2018 original authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package io.micronaut.dbmigration.liquibase.management.endpoint;
 
-import com.fasterxml.jackson.annotation.JsonInclude;
 import io.micronaut.dbmigration.liquibase.LiquibaseConfigurationProperties;
 import io.micronaut.management.endpoint.annotation.Endpoint;
 import io.micronaut.management.endpoint.annotation.Read;
 import io.reactivex.Single;
-import liquibase.changelog.ChangeSet.ExecType;
-import liquibase.changelog.RanChangeSet;
 import liquibase.changelog.StandardChangeLogHistoryService;
 import liquibase.database.Database;
 import liquibase.database.DatabaseFactory;
@@ -15,11 +28,9 @@ import liquibase.exception.DatabaseException;
 
 import javax.sql.DataSource;
 import java.sql.SQLException;
-import java.time.Instant;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
-import java.util.Set;
 import java.util.stream.Collectors;
 
 /**
@@ -89,181 +100,5 @@ public class LiquibaseEndpoint {
         }
 
         return Single.just(reports);
-    }
-
-    /**
-     * A Liquibase change set.
-     */
-    @JsonInclude(JsonInclude.Include.ALWAYS)
-    public static class ChangeSet {
-
-        private final String author;
-
-        private final String changeLog;
-
-        private final String comments;
-
-        private final Set<String> contexts;
-
-        private final Instant dateExecuted;
-
-        private final String deploymentId;
-
-        private final String description;
-
-        private final ExecType execType;
-
-        private final String id;
-
-        private final Set<String> labels;
-
-        private final String checksum;
-
-        private final Integer orderExecuted;
-
-        private final String tag;
-
-        /**
-         * @param ranChangeSet The liquibase changeset
-         */
-        public ChangeSet(RanChangeSet ranChangeSet) {
-            this.author = ranChangeSet.getAuthor();
-            this.changeLog = ranChangeSet.getChangeLog();
-            this.comments = ranChangeSet.getComments();
-            this.contexts = ranChangeSet.getContextExpression().getContexts();
-            this.dateExecuted = Instant.ofEpochMilli(ranChangeSet.getDateExecuted().getTime());
-            this.deploymentId = ranChangeSet.getDeploymentId();
-            this.description = ranChangeSet.getDescription();
-            this.execType = ranChangeSet.getExecType();
-            this.id = ranChangeSet.getId();
-            this.labels = ranChangeSet.getLabels().getLabels();
-            this.checksum = ((ranChangeSet.getLastCheckSum() != null) ? ranChangeSet.getLastCheckSum().toString() : null);
-            this.orderExecuted = ranChangeSet.getOrderExecuted();
-            this.tag = ranChangeSet.getTag();
-        }
-
-        /**
-         * @return The author of the change
-         */
-        public String getAuthor() {
-            return this.author;
-        }
-
-        /**
-         * @return The changelog
-         */
-        public String getChangeLog() {
-            return this.changeLog;
-        }
-
-        /**
-         * @return The comments
-         */
-        public String getComments() {
-            return this.comments;
-        }
-
-        /**
-         * @return The contexts
-         */
-        public Set<String> getContexts() {
-            return this.contexts;
-        }
-
-        /**
-         * @return The execution date for the changeset
-         */
-        public Instant getDateExecuted() {
-            return this.dateExecuted;
-        }
-
-        /**
-         * @return The deployment id
-         */
-        public String getDeploymentId() {
-            return this.deploymentId;
-        }
-
-        /**
-         * @return The descriptions
-         */
-        public String getDescription() {
-            return this.description;
-        }
-
-        /**
-         * @return The {@link ExecType}
-         */
-        public ExecType getExecType() {
-            return this.execType;
-        }
-
-        /**
-         * @return The changeset id
-         */
-        public String getId() {
-            return this.id;
-        }
-
-        /**
-         * @return The labels
-         */
-        public Set<String> getLabels() {
-            return this.labels;
-        }
-
-        /**
-         * @return The checksum
-         */
-        public String getChecksum() {
-            return this.checksum;
-        }
-
-        /**
-         * @return The order in which the migration has been executed
-         */
-        public Integer getOrderExecuted() {
-            return this.orderExecuted;
-        }
-
-        /**
-         * @return The tag
-         */
-        public String getTag() {
-            return this.tag;
-        }
-    }
-
-    /**
-     * Liquibase report for one datasource.
-     */
-    public static class LiquibaseReport {
-
-        private final String name;
-
-        private final List<ChangeSet> changeSets;
-
-        /**
-         * @param name       The name of the data source
-         * @param changeSets The list of changes
-         */
-        public LiquibaseReport(String name, List<ChangeSet> changeSets) {
-            this.name = name;
-            this.changeSets = changeSets;
-        }
-
-        /**
-         * @return The name of the data source
-         */
-        public String getName() {
-            return this.name;
-        }
-
-        /**
-         * @return The list of changesets
-         */
-        public List<ChangeSet> getChangeSets() {
-            return changeSets;
-        }
     }
 }
